task/팀 엔티티 CRUD API 생성 #46 [작업 내용] - 삭제하려는 팀에 사용자가 존재하는 경우, 예외를 발생시켜서 팀에 사용자가 존재한다는 것을 알려주도록 했습니다. - 사용자를 등록할 때, 존재하는 팀 목록을 보여주고 그 팀에 사용자를 넣을 수 있도록 했습니다. - h2database version 업데이트

+++ src/main/java/com/project/voa/service/TeamService.java
@@ -1,20 +1,25 @@
 package com.project.voa.service;
 
 import com.project.voa.domain.Team;
+import com.project.voa.domain.UserInfo;
 import com.project.voa.error.ErrorCodes;
 import com.project.voa.repository.TeamRepository;
 import com.project.voa.dto.TeamModel;
+import com.project.voa.repository.UserInfoRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.StreamSupport;
 
 @Service
 @RequiredArgsConstructor
 public class TeamService {
 	private final TeamRepository teamRepository;
+	private final UserInfoRepository userInfoRepository;
 
 	public TeamModel createTeam(final String name) {
 		Team team = new Team();
@@ -42,7 +47,20 @@ public class TeamService {
 		return new TeamModel(String.valueOf(id), team.getName());
 	}
 
-	public void deleteTeam(long id) {
+	/**
+	 * 삭제할 team에 user가 있는 경우, 에러코드 내려줌.
+	 * @param id
+	 */
+	public void deleteTeam(final long id) {
+		List<UserInfo> users = StreamSupport.stream(userInfoRepository.findAll().spliterator(), false).toList()
+				.stream()
+				.filter(user -> user.getTeam().getId() == id)
+				.toList();
+
+		if (!users.isEmpty()) {
+			throw new DataIntegrityViolationException(ErrorCodes.DO_NOT_DELETE.name());
+		}
+
 		teamRepository.deleteById(id);
 	}
 }
